Remove requests from BFT memory pool when syncing
When a BFT node commits a block, it goes through the transactions in the block and searches whether they exist in the in-memory pool, and if so, it removes them from the pool.

When a follower node syncs blocks from another node, the requests of these blocks still remain in its request pool.
They not only take up memory, but also holds the semaphore resources which throttle needlessly the client.

This commit goes through the requests of a block that is committed through synchronization and removes them as well.
This is needed because the library only sees the last block committed during synchronization and not the entire range.

Signed-off-by: Yacov Manevich <yacov.manevich@ibm.com>
(cherry picked from commit c53daa36ad7e68ccaeb3c707ec14134d079e87c7)